Move static page bodies into their route files
Pure move: each routes/ file previously did nothing but re-export the
page.tsx now in its place. The createFileRoute wrapper is restored in the
next commit, per the no-rename-plus-edit rule.
